requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Document parsing
PyMuPDF>=1.23.0
python-docx>=1.0.0
Pillow>=10.0.0
pytesseract>=0.3.10

# AI / LLM
openai>=1.0.0

# Report generation
reportlab>=4.0.0

# Database
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
httpx>=0.25.0
aiofiles>=23.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
anyio>=4.0.0
